requirements.txt
airsim==1.8.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.1
pyyaml==6.0.1

# 额外建议添加的包
Pillow>=10.0.0        # 图像处理辅助
python-dotenv>=1.0.0  # 环境变量管理（存API Key）